refactor(PyO3): factor code in a macro for python keys
