+++ src/interfaces/pyo3/tests/test_cover_crypt.py
@@ -84,15 +84,21 @@ class TestKeyGeneration(unittest.TestCase):
         self.msk, self.pk = self.cc.generate_master_keys(self.policy)
 
     def test_master_key_serialization(self) -> None:
+        # test deep copy
+        copy_msk = self.msk.deep_copy()
+        self.assertIsInstance(copy_msk, MasterSecretKey)
+
+        copy_pk = self.pk.deep_copy()
+        self.assertIsInstance(copy_pk, PublicKey)
+
+        # test serialization
         msk_bytes = self.msk.to_bytes()
         self.assertIsInstance(MasterSecretKey.from_bytes(msk_bytes), MasterSecretKey)
-
         with self.assertRaises(Exception):
             MasterSecretKey.from_bytes(b'wrong data')
 
         pk_bytes = self.pk.to_bytes()
         self.assertIsInstance(PublicKey.from_bytes(pk_bytes), PublicKey)
-
         with self.assertRaises(Exception):
             PublicKey.from_bytes(b'wrong data')
 
@@ -102,7 +108,11 @@ class TestKeyGeneration(unittest.TestCase):
             'Secrecy::High && (Country::France || Country::Spain)',
             self.policy,
         )
+        # test deep copy
+        copy_usk = self.msk.deep_copy()
+        self.assertIsInstance(copy_usk, MasterSecretKey)
 
+        # test serialization
         usk_bytes = usk.to_bytes()
         self.assertIsInstance(UserSecretKey.from_bytes(usk_bytes), UserSecretKey)
